Automatically mute the microphone during audio playback.
Modified the code to essentially mute the microphone during audio
playback. The external recording app will continue to run and write
to stdout during playback. However, instead of passing that audio
to diatheke, we simply pull it out of the pipe and instead send a
byte of digital silence to diatheke.