Remove audio data size check from ASR streams (C++).
A message with zero bytes of audio is interpreted as EOF by the
server. Decided that it should be the responsibility of SDK users
to decide whether they want to use that feature to end an ASR
stream early, and removed the check from the client wrapper code.
SDK users can check for empty audio data before pushing audio if
they don't want it to terminate the stream.