Add new methods to the Diatheke API.
Added a StartSession method, which will allow a client application
time to set up an event stream before the session's model begins
executing.

Added a SetStory method, which will allow a client application to
change the state of a session to a new story.

Also updated the wording of many comments to be consistent with
the current and near-future state of diathekesvr.