refactor!: Update the C++ SDK to use the v2 API.
This also includes an update to the specific version of gRPC that
CMake will fetch when building the SDK (now set to get v1.33.2).

Also went for a simpler approach to the C++ SDK with much less
wrapping around the gRPC API, so that it is very similar to the
Go and Python SDKs.