feat(api): Allow custom wakewords and add a separate transcribe RPC.
The custom wakeword allows a client application to specify the text
of the wakeword at runtime, which may be useful in situations where
a client app may need to choose one among many possible wakewords.

Also added a separate transcribe function that may be used for
longer transcriptions that are unrelated to Diatheke interactions.
Also added a new TranscribeAction, which will allow us to specify
in a model when a transcription is required in response to an
intent (e.g., "take a note", "send a message"), and will notify
a client application when it should call the transcribe method.

Finally, added some additional comments to the StreamASR method
to indicate that wakeword processing (if handled by Diatheke) will
happen as part of that method.