fix(cmake): Always build gRPC as a static library in C++.
Updated CMake to always build gRPC as a static library, regardless
of whether or not the Diatheke SDK is built as a shared library.
When building a shared library, this results in the gRPC library
being included as part of the libdiatheke_client.so shared library,
instead of also having the many separate *.so libs for gRPC. From
a practical standpoint, it is then much easier to manage the one
libdiatheke_client.so library and link against it in other projects.