docs: Add documentation for the new Transcribe method.
Added descriptions for the new Transcribe method and supporting
Transcribe action. Included examples for Go, Python, and C++.
Also updated the Python StreamASR convenience function example to
use the updated client based method.

Regenerated the protobuf API docs to reflect the latest protobuf
file, and fixed a typo in the protobuf comments that was manifest
in the autogenerated file.