chore: Update version strings to prepare for next release.
Updated strings where necessary in documentation (and a little swift
code) with the version string for the v2.1.0 release of the SDK.

Updated the main README file instructions for tagging a release
to reflect that we no longer commit generated documentation files.

Re-generated go and python grpc code to reflect latest changes to
the protobuf file. All these changes are related to comments, and
should not change any code or behaviors.