fix (cpp): Modify ASRStream implementation to better respond to results.
It turns out that the bool returned by the underlying gRPC stream
Write() method, didn't quite mean what I thought it meant (C++ API
docs aren't very clear on this point). The Write() method will
return false when the stream is closed, but not necessarily when
the result is available. However, the result IS available as soon
as the stream's Finish() method returns. Experiments confirm that
this function can return a good 400-500ms before the Write() method
returns false. So, to improve the notification of when a result is
available, the ASRStream implementation was updated to listen for
a result (via the blocking Finish method) on a separate thread.
Using an atomic bool as a flag, we then let the calling thread know
when a result is available by returning false from the ASRStream's
send methods once the flag is set. This greatly improves the time
it takes for a client to get the ASR result.

Also simplified the creation of the ASRStream by passing in the
client Stub pointer, and better organized the class by storing
internal class data in an opaque pointer (also helps with issues
related to storing a std::thread object in a class).